(debug) analyse shortwave flux at top of model
The shortwave flux at top of model is obtained directly from CAM history's
FSNT and indirectly from FUS and FDS.  A difference is observed in the net
shorwave flux derived from FUS and FDS output by CLIRAD that has been run
straight after RRTMG within the same model time-step.  However, FSNT does
not show the same difference, indicating that it's the way FUS and FDS that
are output that could be the cause.